requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.27,<1.0

SQLAlchemy>=2.0,<3.0
asyncpg>=0.29,<1.0
alembic>=1.13,<2.0

pydantic>=2.6,<3.0
pydantic-settings>=2.2,<3.0
email-validator>=2.1,<3.0

redis>=5.0,<6.0

python-jose[cryptography]>=3.3,<4.0
passlib[bcrypt]>=1.7,<2.0
